Mueve las funciones getCantidad, getCantOtrasSucursales y estaBonificado al service de productos.
