+++ src/app/components/producto-modal/producto-modal.component.html
@@ -26,16 +26,16 @@
               <span>
                 <span class="mr-2 codigo">{{ p.codigo }}</span>
                 <span *ngIf="p.oferta" class="badge badge-success">OFERTA {{ p.porcentajeBonificacionOferta }}%</span>
-                <span *ngIf="!p.oferta && estaBonificado(p)" class="badge badge-success">BONIF. {{ p.porcentajeBonificacionPrecio }}%</span>
+                <span *ngIf="!p.oferta && productosService.estaBonificado(p)" class="badge badge-success">BONIF. {{ p.porcentajeBonificacionPrecio }}%</span>
               </span>
               <span class="mr-2"><strong>{{ p.descripcion }}</strong></span>
-              <span class="mr-2">Stock: <strong>{{ getCantidad(p) }} {{ p.nombreMedida }}</strong> - Otra Suc: {{ getCantOtrasSucursales(p) }} {{ p.nombreMedida }}</span>
+              <span class="mr-2">Stock: <strong>{{ productosService.getCantidad(p) }} {{ p.nombreMedida }}</strong> - Otra Suc: {{ productosService.getCantOtrasSucursales(p) }} {{ p.nombreMedida }}</span>
               <span class="mr-2" *ngIf="p.bulto > 1">Llevando {{ p.bulto }} o más</span>
               <span class="mr-2">
-                <span class="text-nowrap" *ngIf="!((p.oferta || estaBonificado(p)) && p.bulto == 1)">
-                  P.U. <span class="precio" [class.tachado]="(p.oferta || estaBonificado(p)) && p.bulto == 1">{{ p.precioLista | currency:'ARS':'symbol':'1.0-2' }}</span>
+                <span class="text-nowrap" *ngIf="!((p.oferta || productosService.estaBonificado(p)) && p.bulto == 1)">
+                  P.U. <span class="precio" [class.tachado]="(p.oferta || productosService.estaBonificado(p)) && p.bulto == 1">{{ p.precioLista | currency:'ARS':'symbol':'1.0-2' }}</span>
                 </span>
-                <span class="text-nowrap" *ngIf="p.oferta || estaBonificado(p)">
+                <span class="text-nowrap" *ngIf="p.oferta || productosService.estaBonificado(p)">
                   P.U. {{ p.oferta ? "Oferta" : "Bonif."}} <span class="precio">{{ p.precioBonificado | currency:'ARS':'symbol':'1.0-2' }}</span>
                 </span>
               </span>
